# social-media/backend/docs/swagger-auth/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_socialhub_auth-service_pkg_models.Post:
    properties:
      comments_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      edited_at:
        $ref: '#/definitions/sql.NullTime'
      id:
        type: string
      image_url:
        $ref: '#/definitions/sql.NullString'
      is_edited:
        type: boolean
      likes_count:
        type: integer
      shares_count:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
      video_url:
        $ref: '#/definitions/sql.NullString'
      views_count:
        type: integer
    type: object
  github_com_socialhub_auth-service_pkg_models.PostWithDetails:
    properties:
      comments_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      edited_at:
        $ref: '#/definitions/sql.NullTime'
      id:
        type: string
      image_url:
        $ref: '#/definitions/sql.NullString'
      is_edited:
        type: boolean
      likes_count:
        type: integer
      shares_count:
        type: integer
      total_comments:
        type: integer
      total_likes:
        description: Aggregated counts
        type: integer
      total_shares:
        type: integer
      updated_at:
        type: string
      user_avatar:
        type: string
      user_id:
        type: string
      user_is_verified:
        type: boolean
      user_name:
        description: User details
        type: string
      user_reaction:
        description: User's reaction to this post
        type: string
      user_username:
        type: string
      video_url:
        $ref: '#/definitions/sql.NullString'
      views_count:
        type: integer
    type: object
...
  internal_handlers.CreatePostRequest:
    properties:
      content:
        maxLength: 5000
        minLength: 1
        type: string
      image_url:
        type: string
      video_url:
        type: string
    required:
    - content
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  internal_handlers.FeedResponse:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      posts:
        items:
          $ref: '#/definitions/github_com_socialhub_auth-service_pkg_models.PostWithDetails'
        type: array
      total_count:
        type: integer
    type: object
...
  internal_handlers.PostsResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      posts:
        items:
          $ref: '#/definitions/github_com_socialhub_auth-service_pkg_models.Post'
        type: array
    type: object
...
  internal_handlers.UpdatePostRequest:
    properties:
      content:
        maxLength: 5000
        minLength: 1
        type: string
      image_url:
        type: string
      video_url:
        type: string
    type: object
...
  sql.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
  sql.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Authentication service for SocialHub platform
  title: SocialHub Auth API
  version: 1.0.0
paths:
...
  /api/v1/feed:
    get:
      description: Get paginated feed with posts based on feed type
      parameters:
      - default: home
        description: 'Feed type: home, trending, latest'
        in: query
        name: type
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.FeedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get user's feed
      tags:
      - posts
  /api/v1/hashtag/{tag}:
    get:
      description: Get paginated posts with a specific hashtag
      parameters:
      - description: 'Hashtag (without #)'
        in: path
        name: tag
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PostsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get posts by hashtag
      tags:
      - posts
...
  /api/v1/posts:
    post:
      consumes:
      - application/json
      description: Create a new post with content and optional media
      parameters:
      - description: Post creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_socialhub_auth-service_pkg_models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Create a new post
      tags:
      - posts
  /api/v1/posts/{id}:
    delete:
      description: Delete an existing post (only owner can delete)
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Delete a post
      tags:
      - posts
    get:
      description: Get a specific post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_socialhub_auth-service_pkg_models.PostWithDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get a post by ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update an existing post (only owner can update)
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Post update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_socialhub_auth-service_pkg_models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Update a post
      tags:
      - posts
...
  /api/v1/users/{user_id}/posts:
    get:
      description: Get paginated posts by a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PostsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get posts by user
      tags:
      - posts
...
swagger: "2.0"